Move amp to runtime pkg
