+++ flashlight/ext/test/CMakeLists.txt
@@ -3,11 +3,6 @@ cmake_minimum_required(VERSION 3.10)
 set(DIR ${CMAKE_CURRENT_LIST_DIR})
 set(LIBS flashlight)
 
-build_test(
-  SRC ${DIR}/amp/DynamicScalerTest.cpp
-  LIBS ${LIBS}
-)
-
 if(FL_BUILD_CONTRIB)
   build_test(
     SRC ${DIR}/common/SequentialBuilderTest.cpp
