+++ flashlight/pkg/runtime/CMakeLists.txt
@@ -26,13 +26,14 @@ add_library(
   ""
 )
 
+
 target_sources(
   fl_pkg_runtime
   PRIVATE
   ${CMAKE_CURRENT_LIST_DIR}/Runtime.cpp
 )
 
-message("here" ${gflags_LIBRARIES})
+include(${CMAKE_CURRENT_LIST_DIR}/amp/CMakeLists.txt)
 
 target_link_libraries(
   fl_pkg_runtime
@@ -48,6 +49,10 @@ target_include_directories(
   ${GLOG_INCLUDE_DIRS}
   ${gflags_INCLUDE_DIRS}
 )
+
+if (FL_BUILD_TESTS)
+  add_subdirectory(${CMAKE_CURRENT_LIST_DIR}/test)
+endif ()
 
 if (FL_CODE_COVERAGE)
   add_coverage_to_target(TARGET fl_pkg_runtime)
