+++ flashlight/pkg/runtime/amp/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.10)
 
 target_sources(
-  flashlight
+  fl_pkg_runtime
   PRIVATE
   ${CMAKE_CURRENT_LIST_DIR}/DynamicScaler.cpp
   )
